# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IUser": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"familyName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIndexResponse_IUser_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IUser"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.Exclude_keyofIUser.userId__": {
				"properties": {
					"familyName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUser.userId_": {
				"$ref": "#/components/schemas/Pick_IUser.Exclude_keyofIUser.userId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IUserCreate": {
				"$ref": "#/components/schemas/Omit_IUser.userId_"
			},
			"IUpdateResponse": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"rows": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IUserCreate_": {
				"properties": {
					"familyName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IUserUpdate": {
				"$ref": "#/components/schemas/Partial_IUserCreate_"
			},
			"IUserFile": {
				"properties": {
					"fileId": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"storageKey": {
						"type": "string"
					},
					"filename": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					}
				},
				"required": [
					"fileId",
					"userId",
					"storageKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIndexResponse_IUserFile_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IUserFile"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "jwt",
				"in": "header",
				"authorizationUrl": "http://swagger.io/api/oauth/dialog"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/magic": {
			"get": {
				"operationId": "SendMagicLink",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"ok": {
											"type": "boolean"
										}
									},
									"required": [
										"ok"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/login": {
			"get": {
				"operationId": "LoginFromMaginLink",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"redirectTo": {
											"type": "string"
										},
										"renew": {
											"type": "string"
										},
										"access": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"redirectTo",
										"renew",
										"access"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/protected/user": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IUser_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page d'utilisateurs.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer un nouvel utilisateur",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreate"
							}
						}
					}
				}
			}
		},
		"/protected/user/{userId}": {
			"get": {
				"operationId": "ReadUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"description": "Récupérer une utilisateur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un utilisateur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer un utilisateur",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IUser_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page d'utilisateurs.",
				"security": [],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer un nouvel utilisateur",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreate"
							}
						}
					}
				}
			}
		},
		"/user/{userId}": {
			"get": {
				"operationId": "ReadUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"description": "Récupérer une utilisateur avec le ID passé dans le URL",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un utilisateur avec le ID passé dans le URL",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer un utilisateur",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/{userId}/file": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Envoyer un fichier",
				"security": [],
				"parameters": [
					{
						"description": "Le ID de l'utilisateur",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ShowFiles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IUserFile_"
								}
							}
						}
					}
				},
				"description": "Récupérer une liste de fichiers d'un utilisateur",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/{userId}/file/{fileId}": {
			"get": {
				"operationId": "DownloadFile",
				"responses": {
					"200": {
						"description": "Chunked object stream"
					}
				},
				"description": "Récupérer un fichier selon son ID. Le résultat est une série de messages (statut 200) contenant les contenus du fichier.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "fileId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}